+++ packages/test-utils/src/loadFlatEmojiData.ts
@@ -1,12 +1,6 @@
-import { Emoji, flattenEmojiData } from 'emojibase';
+import { Emoji, flattenEmojiData, ShortcodesDataset } from 'emojibase';
 import loadEmojiData from './loadEmojiData';
 
-let DATA: Emoji[] = [];
-
-export default function loadFlatEmojiData(): Emoji[] {
-  if (DATA.length === 0) {
-    DATA = flattenEmojiData(loadEmojiData());
-  }
-
-  return DATA;
+export default function loadFlatEmojiData(shortcodeDatasets?: ShortcodesDataset[]): Emoji[] {
+  return flattenEmojiData(loadEmojiData(), shortcodeDatasets);
 }