+++ packages/test-utils/src/loadMeta.ts
@@ -1,12 +1,6 @@
 import { MetadataDataset } from 'emojibase';
 
-let DATA: MetadataDataset;
-
 export default function loadMeta(): MetadataDataset {
-  if (!DATA) {
-    // eslint-disable-next-line global-require, @typescript-eslint/no-unsafe-assignment
-    DATA = require('../test-meta.json');
-  }
-
-  return DATA;
+  // eslint-disable-next-line global-require
+  return require('../test-meta.json');
 }