+++ packages/test-utils/src/loadShortcodes.ts
@@ -1,12 +1,6 @@
 import { ShortcodesDataset } from 'emojibase';
 
-let DATA: ShortcodesDataset;
-
 export default function loadShortcodes(): ShortcodesDataset {
-  if (!DATA) {
-    // eslint-disable-next-line global-require, @typescript-eslint/no-unsafe-assignment
-    DATA = require('../test-shortcodes.json');
-  }
-
-  return DATA;
+  // eslint-disable-next-line global-require
+  return require('../test-shortcodes.json');
 }